[3958] Fixed the allocation from the large PD pool.

+++ src/lib/dhcpsrv/alloc_engine.cc
@@ -631,8 +631,8 @@ AllocEngine::allocateUnreservedLeases6(ClientContext6& ctx) {
     /// try that number of times at most. It would be useful to that value if
     /// attempts_, specified by the user could override that value (and keep
     /// dynamic if they're set to 0).
-    uint32_t max_attempts = ctx.subnet_->getPoolCapacity(ctx.type_);
-    for (uint32_t i = 0; i < max_attempts; ++i)
+    uint64_t max_attempts = ctx.subnet_->getPoolCapacity(ctx.type_);
+    for (uint64_t i = 0; i < max_attempts; ++i)
     {
         IOAddress candidate = allocator->pickAddress(ctx.subnet_, ctx.duid_, hint);
